Currently, we initialize "error" to zero when it's declared, then we never initialize it to anything else. However, in the case that m_uiotombuf fails, we return error (effectively reporting success).
This appears to be a relic of an older revision of this file, where
"error" used to be doing something useful. (See revision 1.1, where
error is used in a loop with uiomove() instead of using m_uiotomubf).

So instead on unconditionally reporting success in the case there is
a failure in m_uiotombuf, explicitly return ENOBUFS. While we are
here, garbage collect the error variable since it's no longer required.

MFC after:	2 weeks


git-svn-id: svn://svn.freebsd.org/base/head@163986 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f